requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
pyyaml>=6.0
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0
networkx>=3.1
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0 
requests>=2.31.0
certifi>=2023.7.22
urllib3>=2.0.0
pycountry>=24.5.18

streamlit-float>=0.3.0
langchain>=0.0.350
langchain-experimental>=0.0.40
openai>=1.0.0
google-generativeai>=0.3.0
langchain-google-genai>=0.0.8
millify>=0.1.1